Complete component container class and refactor
"order" is replaced by "user command" to respect valve documentation
terminology.
