--- include/order.hpp
@@ -1,20 +1,0 @@
-#ifndef IORDER_HPP_INCLUDED
-#define IORDER_HPP_INCLUDED
-
-#include "components/component.hpp"
-
-namespace trillek {
-namespace order {
-
-    using namespace trillek::component;
-
-    void Execute(id_t id, std::shared_ptr<component::Container>&& order) {
-        if (order->Is<Component::Velocity>()) {
-            Update<Component::Velocity>(id, std::move(order));
-        }
-    }
-
-} // namespace order
-} // namespace trillek
-
-#endif // IORDER_HPP_INCLUDED
